+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support to smart images
+- New defmodule to manipulate the url
+- Method generate instead of HOF to create client
 
 ## [0.1.0] - 2018-05-02
 ### Added